+++ static/app/views/detectors/components/forms/metric/metric.tsx
@@ -1,4 +1,4 @@
-import {Fragment, useContext, useEffect, useMemo} from 'react';
+import {Fragment, useContext, useEffect} from 'react';
 import {useTheme} from '@emotion/react';
 import styled from '@emotion/styled';
 import toNumber from 'lodash/toNumber';
@@ -13,10 +13,7 @@ import type {RadioOption} from 'sentry/components/forms/controls/radioGroup';
 import {NumberField} from 'sentry/components/forms/fields/numberField';
 import {SegmentedRadioField} from 'sentry/components/forms/fields/segmentedRadioField';
 import {SelectField} from 'sentry/components/forms/fields/selectField';
-import {SentryMemberTeamSelectorField} from 'sentry/components/forms/fields/sentryMemberTeamSelectorField';
-import {TextareaField} from 'sentry/components/forms/fields/textareaField';
 import {FormContext} from 'sentry/components/forms/formContext';
-import {useFormField} from 'sentry/components/workflowEngine/form/useFormField';
 import {Container} from 'sentry/components/workflowEngine/ui/container';
 import {FormSection} from 'sentry/components/workflowEngine/ui/formSection';
 import {IconWarning} from 'sentry/icons/iconWarning';
@@ -27,7 +24,6 @@ import {DataConditionType} from 'sentry/types/workflowEngine/dataConditions';
 import type {Detector, MetricDetectorConfig} from 'sentry/types/workflowEngine/detectors';
 import {generateFieldAsString} from 'sentry/utils/discover/fields';
 import {useLocation} from 'sentry/utils/useLocation';
-import {useProjects} from 'sentry/utils/useProjects';
 import {
   AlertRuleSensitivity,
   AlertRuleThresholdType,
@@ -38,6 +34,7 @@ import {
 } from 'sentry/views/detectors/components/details/metric/transactionsDatasetWarning';
 import {useIsMigratedExtrapolation} from 'sentry/views/detectors/components/details/metric/utils/useIsMigratedExtrapolation';
 import {AutomateSection} from 'sentry/views/detectors/components/forms/automateSection';
+import {IssueOwnershipSection} from 'sentry/views/detectors/components/forms/common/issueOwnershipSection';
 import {ProjectEnvironmentSection} from 'sentry/views/detectors/components/forms/common/projectEnvironmentSection';
 import {EditDetectorLayout} from 'sentry/views/detectors/components/forms/editDetectorLayout';
 import type {MetricDetectorFormData} from 'sentry/views/detectors/components/forms/metric/metricFormData';
@@ -636,49 +633,6 @@ function DetectSection({step}: {step?: number}) {
   );
 }
 
-function IssueOwnershipSection({step}: {step?: number}) {
-  const projectId = useFormField<string>('projectId');
-  const {projects} = useProjects();
-  const memberOfProjectSlugs = useMemo(() => {
-    const project = projects.find(p => p.id === projectId);
-    return project ? [project.slug] : undefined;
-  }, [projects, projectId]);
-
-  return (
-    <Container>
-      <FormSection step={step} title={t('Issue Ownership')}>
-        <Stack>
-          <OwnershipField
-            placeholder={t('Select a member or team')}
-            label={t('Assign')}
-            help={t(
-              'Sentry will assign issues detected by this monitor to this individual or team.'
-            )}
-            name="owner"
-            flexibleControlStateSize
-            memberOfProjectSlugs={memberOfProjectSlugs}
-          />
-          <MinHeightTextarea
-            name="description"
-            label={t('Describe')}
-            help={t(
-              'Add any additional context about this monitor for other team members.'
-            )}
-            stacked
-            inline={false}
-            aria-label={t('description')}
-            placeholder={t(
-              'Example monitor description\n\nTo debug follow these steps:\n1. \u2026\n2. \u2026\n3. \u2026'
-            )}
-            rows={6}
-            autosize
-          />
-        </Stack>
-      </FormSection>
-    </Container>
-  );
-}
-
 function TransactionsDatasetWarningListener() {
   const dataset = useMetricDetectorFormField(METRIC_DETECTOR_FORM_FIELDS.dataset);
   if (dataset !== DetectorDataset.TRANSACTIONS) {
@@ -851,16 +805,4 @@ const PriorityLabel = styled('span')`
 const RequiredAsterisk = styled('span')`
   color: ${p => p.theme.tokens.content.danger};
   margin-left: ${p => p.theme.space['2xs']};
-`;
-
-const OwnershipField = styled(SentryMemberTeamSelectorField)`
-  padding: ${p => p.theme.space.lg} 0;
-`;
-
-// Min height helps prevent resize after placeholder is replaced with user input
-const MinHeightTextarea = styled(TextareaField)`
-  padding: ${p => p.theme.space.lg} 0;
-  textarea {
-    min-height: 140px;
-  }
 `;
